Apply GameRule to ChessPiece and others

+++ AmericanChess/source/GameStateManager/State/GSMenu.cpp
@@ -67,7 +67,7 @@ void GSMenu::update(float deltaTime) {
 		m_currentTime += deltaTime;
 		if (m_currentTime < TRANSITION_DURATION*2) {
 			sf::Color color = sf::Color::Black;
-			color.a = 255 * (1 - (m_currentTime / (TRANSITION_DURATION*2)));
+			color.a = (int)(255 * (1 - (m_currentTime / (TRANSITION_DURATION * 2))));
 			m_blackScreen->setFillColor(color);
 		} else {
 			isPerformTransition = false;
